2022_07_11_Zapata_H2_trotter_icm/requirements.txt
attrs==20.3.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.1.0
cirq==0.13.1
cirq-aqt==0.13.1
cirq-core==0.13.1
cirq-google==0.13.1
cirq-ionq==0.13.1
cirq-pasqal==0.13.1
cirq-rigetti==0.13.1
cirq-web==0.13.1
cma==3.2.2
cmake==3.22.5
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.5.1
duet==0.2.7
fonttools==4.33.3
gcc7==0.0.7
google-api-core==1.31.6
google-auth==1.35.0
googleapis-common-protos==1.56.3
grpcio==1.47.0
h11==0.9.0
h5py==3.7.0
httpcore==0.11.1
httpx==0.15.5
idna==2.10
immutables==0.6
iniconfig==1.1.1
iso8601==0.1.16
joblib==1.1.0
kiwisolver==1.4.3
lark==0.11.3
matplotlib==3.5.2
mpmath==1.2.1
msgpack==0.6.2
networkx==2.8.4
ntlm-auth==1.5.0
numpy==1.23.0
openfermion==1.3.0
openfermionpyscf==0.5
orquestra-cirq==0.3.0
orquestra-core==0.3.0
orquestra-forest==0.3.0
orquestra-opt==0.3.0
orquestra-qiskit==0.3.0
orquestra-quantum==0.3.0
orquestra-qulacs==0.2.0
orquestra-vqa==0.3.0
overrides==6.1.0
packaging==21.3
pandas==1.4.3
pbr==5.9.0
Pillow==9.1.1
pluggy==1.0.0
ply==3.11
protobuf==3.20.1
psutil==5.9.1
PubChemPy==1.0.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyJWT==1.7.1
pyparsing==3.0.9
pyquil==3.0.1
pyscf==2.0.1
pytest==7.1.2
python-constraint==1.4.0
python-dateutil==2.8.2
python-rapidjson==1.6
pytz==2022.1
pyzmq==23.2.0
qcs-api-client==0.8.0
qiskit==0.34.0
qiskit-aer==0.10.1
qiskit-ibmq-provider==0.18.3
qiskit-ignis==0.7.0
qiskit-terra==0.19.1
Qulacs==0.2.0
requests==2.28.1
requests-ntlm==1.1.0
retry==0.9.2
retrying==1.3.3
retworkx==0.11.0
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.10.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stevedore==3.5.0
symengine==0.9.2
sympy==1.9
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
tweedledum==1.1.1
typing-utils==0.1.0
typing_extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.3
